added unique carts for users and different prices for stores, in process of adjusting rest of code to work with changes.

+++ capstone/src/main/java/com/example/capstone/repository/StoreRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.capstone.entity.StoreEntity;
 
 @Repository
-public interface StoreRepository extends JpaRepository<StoreEntity, Integer> {
+public interface StoreRepository extends JpaRepository<StoreEntity, Long> {
 	
 	Optional<StoreEntity> findByName(String name);
 }
